Extract document execution sessions
CellExecutions mixes global session lookup with the document-scoped state and
presentation worker it constructs.

Move document execution behind an effectfully constructed, document-scoped
class while retaining the existing registry and ordering behavior. Publish
staleness as changed cells so the session does not own a second read projection.
